refactor: 커스텀 훅 분리, 수치 상수화 처리 - 계산 부분 useSavingsCalculator 분리 - 수치 부분 상수화 공통 관리
