+++ src/constants/index.ts
@@ -1,1 +1,3 @@
-export {};
+export const SAVINGS_INTEREST_FACTOR = 0.5;
+export const SAVINGS_ROUND_UNIT = 1000;
+export const RATE_PERCENT = 100;
